+++ src/utils/Bind.ts
@@ -4,12 +4,17 @@ import { log } from "../../build/macroses/log.macro";
 
 /**
  * Bind decorator wrappers factory for methods binding.
+ *
+ * It is not obvious for javascript users that changing prototype descriptor of base class
+ * will affect only one instance and all instances after its change.
+ * This util is not intended to be used for all user-cases, only for library.
+ * I would rather tell people to use another approach with instance binding than running into issues
+ * with incorrect ES inheritance after runtime modification of base class.
  */
 function createBoundDescriptor <T>(from: TypedPropertyDescriptor<T>, property: PropertyKey): PropertyDescriptor {
   log.info("Created bound descriptor for:", property);
 
   // Todo: Wait for autobind merge with fix of shared callbacks issue and other.
-  let fn: T = from.value as T;
   let definingProperty: boolean = false;
 
   return ({
@@ -19,32 +24,27 @@ function createBoundDescriptor <T>(from: TypedPropertyDescriptor<T>, property: P
         definingProperty ||
         // this === target.prototype || - will it fire? Check parent prototypes?
         Object.prototype.hasOwnProperty.call(this, property) ||
-        typeof fn !== "function"
+        typeof from.value !== "function"
       ) {
-        return fn;
+        return from.value as any;
       }
 
-      const bound: T = (from as any).value.bind(this);
+      const bound: T = from.value.bind(this);
 
       definingProperty = true;
 
       Object.defineProperty(this, property, {
         configurable: true,
-        get() {
-          return bound;
-        },
-        set(value) {
-          fn = value;
-          delete this[property];
-        }
+        writable: false,
+        value: bound
       });
 
       definingProperty = false;
 
       return bound;
     },
-    set(value: any) {
-      fn = value;
+    set() {
+      throw new Error("Direct runtime modification of decorated in a declarative way method is not allowed.");
     }
   });
 }